Update README.md
Small tweaks to hopefully help other newbies like myself - use default frigate port, and add not on compreface service to use.

Another suggestion would be not using port 3000 as this is the standard grafana port (and I think that it's contention for other apps too) - anything to avoid the small obstacles that can make installing this as easy as possible.  Great app - thanks for you efforts.