Add event type to notification system
This commit introduces the event type to the notification system. The 'types' array has been added to the 'notify' object in the defaults.js file. This array includes 'snapshot', 'latest', 'mqtt', 'frigate', and 'manual' as possible types.

The event type is also added to the output object in the recognize.controller.js file.

In the schemas/index.js file, the 'types' property has been added to the 'notify' object schema.

Finally, in the notify/actions/index.js file, a check has been added to ensure that the output type is included in the NOTIFY.TYPES array before proceeding with the notification. If the output type is not included, a debug message is logged.