Update SQLite and better-sqlite3 versions
This commit updates the SQLite amalgamation version from 3420000 to 3430000 in the Dockerfile. It also updates the better-sqlite3 version from 8.4.0 to 8.6.0 in the Dockerfile and from 8.2.0 to 8.6.0 in the package.json file. These updates are necessary to keep the project dependencies up-to-date and ensure compatibility with the latest features and improvements.