Update IP filter condition in app.js
Modified the condition for applying the IP filter in app.js. Previously, the filter was applied if the HA_ADDON environment variable was set. Now, it will only be applied if both the HA_ADDON and IPFILTER environment variables are set to 'true'. This change provides more flexibility in controlling when the IP filter is applied.