Fix Top Menu Bug and Update Version
This commit addresses a bug in the top menu of the application. The issue has been resolved and the functionality of the top menu should now be restored.

In addition, the version of the application has been updated in both the main package.json and the api/package.json files. The new version is 1.13.11.6.1. This version update reflects the bug fix and any other minor changes that have been made.