Update Dockerfile to use newer version of oven/bun
This commit updates the Dockerfile to use a newer version of oven/bun (1.0.11 instead of 1.0.7). It also removes the 'bun upgrade --canary' command, which is no longer necessary with the updated version of oven/bun. This change should improve the build process and potentially fix any issues caused by the older version of oven/bun.