Add SQL logging and improve error handling
This commit introduces SQL logging in the application by adding a new 'sql' property in the 'logs' object in the defaults.js file. This will allow for better tracking and debugging of SQL queries in the application.

Additionally, the error handling in the recognize.controller.js file has been improved. Instead of using console.log to print the error, console.debug is used to provide more detailed information about the error.

Finally, the db.util.js file has been updated to use the new 'sql' property from the 'logs' object when creating a new Database instance. This will enable or disable SQL logging based on the value of the 'sql' property.