Update to version 1.13.11.7 with improvements
This commit updates the application to version 1.13.11.7. It reintroduces support for the arm/v7 architecture and removes unused code to reduce the size of the Docker container. Performance optimizations have been implemented for better efficiency. Additionally, an option to trace SQL queries has been added for better debugging and tracking. The version changes have been reflected in the package.json files.