Add multi-platform binaries for double-take
This commit adds multi-platform support for the double-take application. It includes changes to the build.yml file to create docker containers for different platforms (linux/arm/v7, linux/arm64, linux/amd64) and export their file systems. These file systems are then compressed and used to build Go binaries for each platform. The built binaries are uploaded as artifacts.

Additionally, a new Go module is added to the .build/bundle directory, which includes the main.go file that handles the execution of the double-take application in different environments. This includes functions for mounting file systems, handling system calls, and managing process attributes.

The commit also updates the Docker meta step in the GitHub workflow to correctly set the Docker tag based on the branch name.

Lastly, the commit includes a new uninstall command in the CLI app to remove the application.