Add Playwright Tests and Update Dependencies
This commit introduces Playwright tests for the application. A new GitHub workflow has been added to run these tests on push and pull request events on the main and master branches. The tests are configured to run on the latest Ubuntu environment and use Node version 18.

The Playwright configuration file and test specifications have also been added. The tests are designed to run on Chromium, Firefox, and Webkit browsers.

In addition, this commit updates the .gitignore file to ignore test results and Playwright cache.