Use environment variable for frontend path
This change updates the static file serving logic in app.js to use a new environment variable, FRONTEND, if it's set. This allows for more flexibility in specifying the path to the frontend assets, which can be particularly useful in different deployment scenarios. The previous setup was limited to a hardcoded path with a conditional for the 'development' environment. Now, the path can be dynamically set through the environment, which can simplify configuration management and deployment processes.