Add MQTT expiration configuration (#181)
This commit introduces a centralized expiration setting for MQTT messages by adding an `expire_after` field to the default configuration. The `mqtt.util.js` file has been updated to reference this new default value instead of using a hardcoded expiration time. This change allows for easier adjustments of the expiration time for MQTT messages in the future.