docs(README): include pose plugin in face_plugins example
This commit updates the README.md to include the 'pose' plugin in the list of example plugins for face detection. This change is intended to guide users on how to extend functionality with additional plugins.

feat(db-util): extend database schema with pose data

In db.util.js, a new field for storing pose information is added to the database schema. The migration function is updated to extract pose data and save it into the new column. This enhancement is part of an effort to handle additional attributes related to facial recognition features.